python-shop-detector/requirements.txt
opencv-python>=4.9.0.80
# Core ML stack
torch>=2.3,<3.0
torchvision>=0.18,<1.0
transformers>=4.43
# Imaging
pillow>=10.3
# Offline text-to-speech
pyttsx3>=2.90
# Web API (local, optional)
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
# Utils
numpy>=1.26
